renamed main file puppy_app.rb to avoid confusion
